+++ src/features/home/slideshow/Slideshow.tsx
@@ -1,57 +1,75 @@
-import React from 'react';
-import type { ReactNode } from 'react';
-
-import useSlideshow from './useSlideshow';
-import Previous from './components/button/Previous';
-import Next from './components/button/Next';
-import Dot from './components/button/Dot';
+import useSlideshow from './hooks/useSlideshow';
+import useCarousel from './hooks/useCarousel';
+import SlideshowLoading from './components/Loading';
+import SlideshowError from './components/Error';
+import SlideshowNavigation from './components/navigation';
+import SlideshowSlide from './components/slide/Slide';
+import SlideshowDots from './components/dots';
 
 interface SlideshowProps {
-  children: ReactNode;
-  height?: number;
+  limit?: number;
+  category?: string;
+  height?: 'standard' | 'tall';
+  button?: boolean;
+  layout?: 'overlay-top' | 'overlay-center' | 'overlay-bottom' | 'split';
+  align?: 'left' | 'center' | 'right';
+  excerptLength?: number;
   loop?: boolean;
   autoplay?: boolean;
   multiSlide?: 2 | 3 | 4;
   navPosition?: 'inside' | 'outside';
 }
 
-interface SlideChildProps {
-  multiSlide?: 2 | 3 | 4;
-  heightClasses?: string;
-}
+export default function Slideshow({
+  limit = 5,
+  category,
+  height = 'standard',
+  button = true,
+  layout = 'overlay-center',
+  align = 'center',
+  excerptLength = 30,
+  loop = true,
+  autoplay = false,
+  multiSlide,
+  navPosition = 'outside',
+}: SlideshowProps) {
+  const slideshowQuery = useSlideshow({ limit, category });
 
-export function Slideshow({ children, height = 500, loop = true, autoplay = false, multiSlide, navPosition = 'outside' }: SlideshowProps) {
-  const { emblaRef, scrollSnaps, selectedIndex, scrollPrev, scrollNext, scrollTo } = useSlideshow(loop, autoplay);
+  const { emblaRef, scrollSnaps, selectedIndex, scrollPrev, scrollNext, scrollTo } = useCarousel(loop, autoplay);
 
-  const heightClasses = height === 500 ? 'md:h-125' : height === 600 ? 'md:h-150' : 'md:h-125';
+  if (slideshowQuery.isPending) return <SlideshowLoading height={height} multiSlide={multiSlide} navPosition={navPosition} />;
+
+  if (slideshowQuery.isSuccess && (!slideshowQuery.data || slideshowQuery.data.length === 0)) return <SlideshowError />;
+
+  const slides = slideshowQuery.data;
+
+  const heightClass = height === 'standard' ? 'md:h-125' : height === 'tall' ? 'md:h-150' : 'md:h-125';
   const multiSlideClass = multiSlide ? '-mr-10' : '';
 
-  // clone and inject props into each child
-  const childrenWithProps = React.Children.map(children, (child) => {
-    if (!React.isValidElement<SlideChildProps>(child)) {
-      return child;
-    }
-
-    return React.cloneElement(child, {
-      multiSlide,
-      heightClasses,
-    });
-  });
-
   return (
-    <div className="embla-carousel relative">
-      <div className={`embla-navigation static left-0 h-full w-full md:absolute md:top-0 md:bottom-0 ${heightClasses}`}>
-        <Previous scrollPrev={scrollPrev} navPosition={navPosition} />
-        <Next scrollNext={scrollNext} navPosition={navPosition} />
-      </div>
-      <div className="embla overflow-hidden rounded-sm" ref={emblaRef}>
-        <div className={`embla__container ${multiSlideClass} flex ${heightClasses}`}>{childrenWithProps}</div>
-      </div>
-      <div className="embla__dots mt-10 flex items-center justify-center gap-x-4">
-        {scrollSnaps.map((_, index) => (
-          <Dot key={index} selected={index === selectedIndex} onClick={() => scrollTo(index)} />
-        ))}
-      </div>
-    </div>
+    <>
+      {slideshowQuery.isSuccess && slides && slides.length > 0 && (
+        <div className="embla-carousel relative" aria-label="Carousel">
+          <SlideshowNavigation scrollPrev={scrollPrev} scrollNext={scrollNext} navPosition={navPosition} heightClass={heightClass} />
+          <div className="embla overflow-hidden rounded-md" ref={emblaRef}>
+            <div className={`embla__container ${multiSlideClass} flex ${heightClass}`}>
+              {slides.map((slide) => (
+                <SlideshowSlide
+                  key={slide.id}
+                  slide={slide}
+                  button={button}
+                  layout={layout}
+                  align={align}
+                  excerptLength={excerptLength}
+                  multiSlide={multiSlide}
+                  heightClass={heightClass}
+                />
+              ))}
+            </div>
+          </div>
+          <SlideshowDots scrollSnaps={scrollSnaps} selectedIndex={selectedIndex} scrollTo={scrollTo} />
+        </div>
+      )}
+    </>
   );
 }